Bug Fixes
I fixed the audio bug on levels 1-4 by switching lines in the script around and turning off the mode play on wake. I changed some camera views and editted the level select menu so the text stays on the screen. I attempted to add glow to level 4 but decided to wait.